analysis/requirements.txt
pyyaml
matplotlib
numpy
scipy
pandas
seaborn
ipykernel